Moved Buckets and Items components to the containers folder
